requirements.txt
asgiref==3.7.2
async-timeout==4.0.3
Django==5.0
django-cors-headers==4.3.1
django-crispy-forms==1.14.0
django-crispy-forms-registration==0.1.3
django-debug-toolbar==4.2.0
django-registration==2.2
django-seed==0.3.1
djangorestframework==3.14.0
Faker==22.0.0
psycopg2-binary==2.9.9
python-dateutil==2.8.2
pytz==2023.3.post1
redis==5.0.1
six==1.16.0
sqlparse==0.4.4
toposort==1.10
typing_extensions==4.9.0
